refact - furnitureDao getAll()

+++ init.sql
@@ -58,10 +58,10 @@ CREATE TABLE projet.visits (
 CREATE TABLE projet.furnitures (
 	furniture_id SERIAL PRIMARY KEY,
 	furniture_title VARCHAR(100) NOT NULL,
-	purchase_price INTEGER NOT NULL,
+	purchase_price FLOAT NOT NULL,
 	furniture_date_collection TIMESTAMP NULL,
-	selling_price INTEGER NOT NULL,
-	special_sale_price INTEGER NOT NULL,
+	selling_price FLOAT NOT NULL,
+	special_sale_price FLOAT NOT NULL,
 	state_furniture state_furniture,
 	deposit_date TIMESTAMP NULL,
 	date_of_sale TIMESTAMP NULL,
